+++ poptorch/source/popart_canonicalization/PopartCanonicalizationUtils.hpp
@@ -35,6 +35,9 @@ SymbolHandler getHandler(torch::jit::NodeKind kind);
 // Returns true if all inputs are Bools
 bool allInputsBool(torch::jit::Node *node, int ignore_input = -1);
 
+// Returns true if all inputs are Bools
+bool allInputsInteger(torch::jit::Node *node, int ignore_input = -1);
+
 // Get the tensor shape as a vector of ints.
 std::vector<std::int64_t> shapeFromTensor(torch::jit::Value *value);